[major] Tidy up package modules
