+++ src/mas/devops/mas/__init__.py
@@ -1,0 +1,15 @@
+from apps import (  # noqa: F401
+    verifyAppInstance,
+    getAppsSubscriptionChannel
+)
+
+from suite import (  # noqa: F401
+    isAirgapInstall,
+    getDefaultStorageClasses,
+    getCurrentCatalog,
+    listMasInstances,
+    getWorkspaceId,
+    verifyMasInstance,
+    getMasChannel,
+    updateIBMEntitlementKey,
+)
